refactor(hook): useIsLg를 mediaQuery를 받을 수 있는 형태로 리팩토링
Resolves: #212
